readme for v3 and small fixes
