+++ v3/io.go
@@ -24,3 +24,24 @@ func (r *Reader) Close() (err error) {
 	}
 	return
 }
+
+// Writer it's a wrapper for given writer, but with progress handle
+type Writer struct {
+	io.Writer
+	bar *ProgressBar
+}
+
+// Write writes bytes to wrapped writer and add amount of bytes to progress bar
+func (r *Writer) Write(p []byte) (n int, err error) {
+	n, err = r.Writer.Write(p)
+	r.bar.Add(n)
+	return
+}
+
+// Close the wrapped reader when it implements io.Closer
+func (r *Writer) Close() (err error) {
+	if closer, ok := r.Writer.(io.Closer); ok {
+		return closer.Close()
+	}
+	return
+}
